refactor: 단위 테스트 작성에 따른 TestFixture 이동 및 이름 변경
